Switch to example use case
Imagines a service that needs to listen for messages on a few different
topics in order to take action. Starts by creating the RabbitMQ
connection using Messenger()'s __init__, then sets up a separate
consumer for each topic pattern that needs matched. The consumers each
call a handler that executes logic defined in a separate module, then
send a follow-up message, if desired.