Add integration tests on API endpoints.
Simple tests just asserting that the returned value (or lack thereof) is
what it should be when interacting with the server as an external actor.

Some small modifications were made to the application deserialization
(handles decoding bytes & json now) & the API for defining a
RouteHandler's arguments was simplified by dropping the keyword argument
requirement.

Finally, a dev stack was configured via docker-compose to allow for
quickly spinning up an instance of this service, with a volume bound to
the local ./app/ directory for easy development & a broker for the
service to communicate with.